Allow dependencies between function pointers in GameEvent objects
Reinforcement cards should prompt the casting player about which side
the resources will go to on *cast* rather than on resolution. This way
if a player respond with a reinforcement of their own they know what the updated score would be.
Also, if reinforcements could be countered the decision to counter would depend on which side
the reinforcements are added to.

Since the side the resources will go to will not be determined until the casting lambda is called
and since the resolution lamdbda depends on this result, we need to ensure a reference to *the same*
GameEvent object is given to both lambdas. Hence we needed to expand the scope of GameEvent objects
as they're created. For the empty stack case this wasn't difficult but for the non-empty case this
scope change required some more work.

Also, the Human's power can only be applied to its own side. Prior to this patch the player was
able to choose sides, which was incorrect.